refactor: move NewGameService to a shared services directory and update import paths
