+++ frontend/src/app/services/new-game.service.ts
